Fix stripPath to stop it URL-decoding the path
Instead of using req.params[0] it now strips the originalUrl
up to the position of the first wildcard placeholder.

For example, when matching /test/* with /test/foo%2Fbar
    - Before this change, stripped URL is foo/bar
    - After this change, stripped URL is /foo%2Fbar

This also fixes a situation that could cause the incorrect path
to be proxied onwards if there are multiple wildcards.

For example, when matching /test/:foo/then/:bar with /test/foovar/then/barval
    - Before this change, stripped URL is fooval
    - After this change, stripped URL is /fooval/then/barval